[Fix] Platform compilation standard.

+++ Plugins/SPlanner/Source/Public/SPlanner/AI/Planner/SP_AIPlannerFlags.h
@@ -8,25 +8,25 @@ UENUM(BlueprintType, Category = "SPlanner|Planner|AI")
 enum class ESP_AIPlannerFlags : uint8
 {
 	/** Position has changed. */
-	PF_LocationDirty	= ESP_PlannerFlags::PF_Max		UMETA(DisplayName = "LocationDirty"),
+	PF_LocationDirty	= static_cast<uint8>(ESP_PlannerFlags::PF_Max)	UMETA(DisplayName = "LocationDirty"),
 
 	/** Rotation has changed. */
-	PF_RotationDirty									UMETA(DisplayName = "RotationDirty"),
+	PF_RotationDirty													UMETA(DisplayName = "RotationDirty"),
 
 	/** Target has changed. */
-	PF_TargetDirty										UMETA(DisplayName = "TargetDirty"),
+	PF_TargetDirty														UMETA(DisplayName = "TargetDirty"),
 
 	/** The target will a Position. */
-	PF_TargetPosition									UMETA(DisplayName = "TargetPosition"),
+	PF_TargetPosition													UMETA(DisplayName = "TargetPosition"),
 
 	/** The target will be an Actor. */
-	PF_TargetActor										UMETA(DisplayName = "TargetActor"),
+	PF_TargetActor														UMETA(DisplayName = "TargetActor"),
 
 	/** The target will be a Player. */
-	PF_TargetPlayer										UMETA(DisplayName = "TargetPlayer"),
+	PF_TargetPlayer														UMETA(DisplayName = "TargetPlayer"),
 
 	/** The target will be a POI. */
-	PF_TargetPOI										UMETA(DisplayName = "TargetPOI"),
+	PF_TargetPOI														UMETA(DisplayName = "TargetPOI"),
 
 	/**
 	*	Max value of enum.
